Add auto-migration support of adding columns
Fixes #129

We support adding columns to existing integrations without having to bump the version;
changing types, or removing/renaming columns, is not supported and should bump the version
or must be handled out-of-band (like deleting the integration then backfilling).
To figure out what columns we need to add, we can check what are currently defined,
check what exists, and add denormalized columns and indices for those that are missing.