Overhaul how webhook responses are handled
Because Warden overrides the 401,
we don't get useful diagnostic information.
This adds a 'reason' header to the rejection.

It also returns an object instead of
the Rack triple, in case we need future adjustment.

Finally, and importantly, it allows disabling of verification
of individual integrations. In some cases,
we may need to backfill from logged webhooks,
but the signature has a timestamp,
and we want to bypass verification.