Restructure DB schemas to be declarative
As part of writing to a different database (see #330)
we need to translate our schema operations
to be more easily targetable.

This creates a new `DBAdapter` namespace,
which models tables/columns/indices,
and an adapter base class that can be used
to convert those strong types into strings
that can be executed against a target database.

The widest change is that we have to use
'generic types' when integrations describe their schema,
so every column declaration changed.

One nice outcome of this is that schema operations
are now decoupled from implementation,
and things like enrichment table schema can be created
in exactly the same way as normal table schema.

This change does not attempt to be comphrensive-
additional modifications will need to be made for Snowflake
(for example, we currently hardcode the adapter to use),
and other more exotic databases may require
more extensive abstraction of the adapters
(for example, we assume the standard SQL
of ';' between statements is valid).