Redo connection cache
Fixes https://github.com/lithictech/webhookdb-api/issues/366

Keep a dynamic cache of open database connections.
Very similar in behavior to Sequel::DATABASES,
but we need to prune connections that have not been inactive for a while.

When `borrow` is called, either a new connection is made,
or an existing one used, for that URL. The connection is yield to the block.

Then, after the block is called,
if 'prune_interval' has elapsed since the last prune,
prune all connections with 0 current connections,
_other than the connection just used_.
Because this connection was just used,
we assume it will be used again soon.

The idea here is that:

- We cannot connect to the DB statically; each org can have its own DB,
  so storing it statically would increase DB connections to the the number of orgs in the database.
- So we replace the organization/synchronization done in Sequel::DATABASES with ConnectionCache.
- Any number of worker threads need access to the same DB; rather than connecting inline,
  which is very slow, all DB connections for an org (or across orgs if not in database isolation)
  can share connections via ConnectionCache.
- In single-org/db environments, the active organization will always always be the same,
  so the connection is never returned.
- In multi-org/db environments, busy orgs will likely stay busy. But a reconnect isn't the end
  of the world.
- It seems more efficient to be pessimistic about future use, and prune anything with 0 connections,
  rather than optimistic, and use an LRU or something similar, since the connections are somewhat
  expensive resources to keep open for now reason. That said, we could switch this out for an LRU
  it the pessimistic pruning results in many reconnections. It would also be reasonable to increase
  the prune interval to avoid disconnecting as frequently.