Durable jobs support (#384)
Fixes https://github.com/lithictech/webhookdb-api/issues/377

Durable jobs keep track of the job in a database, similar to DelayedJob,
so that if Sidekiq loses the job (because Redis crashes, or the worker crashes),
it will be restarted from the data in the database.

The way it works at a high level is:

- Connections to a series of database servers are held.
  These servers act as the 'durable stores' for Redis.
- Directly after a job is enqueued (perform_async, etc),
  a row is written into the first available durable store database.
  Every row records when it should be considered "dead";
  that is, after this time,
  DurableJob is free to re-enqueue a new instance of the job,
  as explained below.
  This is known as the "assume dead at" time; the difference between when a job is enqueued/runs,
  and when it can be assumed dead, is known as the "heartbeat extension".
- Whenever the job runs (Worker#perform), it takes a lock on the durable store row,
  and updates assume_dead_at to be "now plus heartbeat_extension".
  This is true when the job runs the first time, but also during any retry.
- Any long-running jobs should be sure to call DurableJob.heartbeat
  to extend the assume_dead_at, so we don't attempt to enqueue another instance
  of the job (actually we probably won't end up with duplicate jobs,
  but it's a good optimization).
- If the job succeeds, the row is deleted from the durable store.
- If the job errors, assume_dead_at is updated, and the row remains in the durable store.

That is the behavior of the durable jobs themselves.
The other key piece here is a poller. The poller must use a separate periodic mechanism,
like sidekiq-cron or whatever. Some number of minutes, `Amigo::DurableJob.poll_jobs` must be called.
`poll_jobs` does the following at a high level (see the source for more details):

- Look through each durable store database.
- For each job with an assume_dead_at in the past, we need to check whether we should re-enqueue it.
- If the job is currently processing in a queue, we no-op.
  We warn that the job should have a longer heartbeat_extension,
  or something should call DurableJob.heartbeat.
- If the job is currently in the retry set, we update the assume_dead_at of the row
  so it's after the time the job will be retried. That way we won't try and process
  the job again until after it's been retried.
- If the job is in the DeadSet, we delete the row since we should nto re-enqueue it.
- If the job cannot be found in any of these places,
  we re-enqueue the job.

Note that DurableJob is subject to race conditions,
and a job can be enqueued and then run multiple times.
This is an expected part of Sidekiq- your jobs should already
be idempotent so this race should not be an issue.
There are (hopefully) no situations where the race condition
will result in jobs being lost, just processed multiple times.