Add SemaphoreBackoffJob, use for webhook processing
Semaphore backoff jobs can reschedule themselves to happen at a later time
if there is too high a contention on a semaphore.
Ie, if there are too many jobs with the same key,
they start to reschedule themselves for the future.

This is useful when a certain job (or job for a certain target)
can be slow so should not consume all available resources.

In general, you should not use semaphore backoff jobs for singletons,
as the guarantees are not strong enough.
It is useful for many rapid jobs.

Implementers must override the following methods:

- `semaphore_key` must return the Redis key to use as the semaphore.
  This may be something like "sbj-user-1" to limit jobs for user 1.
- `semaphore_size` is the number of concurrent jobs.
  Returning 5 would mean at most 5 jobs could be running at a time.

And may override the following methods:

- `semaphore_backoff` is called to know when to schedule the backoff retry.
  It is called with with the number of concurrent jobs,
  and should return the number of seconds after which to schedule a job.
  Ie with a semaphore size of 5, it may be called with 6; if it returns 60,
  the job would be retried in a minute.
- `semaphore_expiry` should return the TTL of the semaphore key.
  Defaults to 30 seconds. See below for key expiry and negative semaphore value details.
- `before_perform` is called before calling the `perform` method.
  This is required so that implementers can set worker state, based on job arguments,
  that can be used for calculating the semaphore key.

Note that we give the semaphore key an expiry. This is to avoid situation where
jobs are killed, the decrement is not done, and the counter increases to the point we
have fewer than the expected number of jobs running.

This does mean that, when a job runs longer than the semaphore expiry,
another worker can be started, which would increment the counter back to 1.
When the original job ends, the counter would be 0; then when the new job ends,
the counter would be -1. To avoid negative counters (which create the same issue
around missing decrements), if we ever detect a negative 'jobs running',
we warn and remove the key entirely.