Support adding indices to existing columns
Fixes https://github.com/lithictech/webhookdb-api/issues/443

We were only adding indices to newly-added columns,
not columns that were `index: false` that become `index: true`.
We now look for the actual indices that exist in the database.

As part of this, I found that when we were adding an index
to an existing table, the index name was wrong
(not matching the create call).