Release script needs set -e
Without `set -e` (exit on error),
the script would fail the Rake task,
but then continue on, passing the `if` check
about integration tests and exiting 0.

So Heroku would see the release as successful,
even though it failed. We saw the traceback in the logs
so assumed failure, but the exit code was still 0.