Ensuring schemas must handle enrichments
Org.migrate_replication_schemas was not taking into account
whether we want to ensure the `enrichment` column exists,
so if we change `_store_enrichment_body?` to `true`,
we don't create the column and will then error.

This tries to unify the logic a bit around the enrichment column,
ensuring it's `nil` if we don't intend to store the body,
and creating a new 'storable columns' helper
that tries to represent what columns need storage;
I'm not sure it's exactly right conceptually
but it's a start.

Fixes https://sentry.io/organizations/lithic-technology/issues/3546344213/