Migration replication wasn't happening on release
Because the async subscribers weren't set up
(`Webhookdb.register_subscribers`), the calls to `publish`
made in the release phase were going nowhere.

Rather than register subscribers, I opted to use regular Sidekiq jobs.
So we `perform_async` instead of `publish`.

There isn't much benefit in this case to using pubsub since we are
aiming for a very specific job to be done behind the scenes;
pubsub is more about when the publisher doesn't know about
what actions should be taken.