Logged webhooks capture request path and method
Rather than hard-coding it into the logged webhook retry,
it should be captured at log time.

This is not a safe migration in terms of errors,
since we add a non-nullable column.
But I'm okay with the downtime implications while the DB is updated
and the app is not yet.