Overhaul job retry system
Fixes https://github.com/lithictech/webhookdb-api/issues/490

Previously, anything could raise a `Retry` error and
the job would be retried.

However we also want some jobs to die,
not be retried.

Jobs can now raise an `Amigo::Die`, `Amigo::Retry,`
or `Amigo::RetryOrDie`, depending on the desired behavior
(retry or die will happen after a certain number of attempts).

In the end this basically looks like re-writing the
Sidekiq retry mechanics so they can be driven by
the client code, not the job fields.
This is usually not useful, since the jobs and purpose-built
for a certain workload that won't vary that much.

But for WHDB, the ProcessWebhook job in particular
has no particular behavior. It can be everything from a pure function,
to taking several minutes. So we need a custom retry mechanism.

The alternative would be job classes that vary per-replicator,
since the failure modes are pretty similar for a single replicator.
I'm hesitant to make that change, however- it will introduce
a whole new set of complexity. For now, I think it's fine to
take this complexity, but if we find that it is unreliable
or buggy, we should look at jobs per-repliocator.