Body-optional webhooks, and svcint sequences
Up until now, we've assumed everything about a webhook
would be expressed via its body. This was slightly changed,
so that `upsert_webhook` can take info about headers, request method,
and request path, but it didn't make it all the way to things like
the column defaulters and converters.

This adds a few things to support things like
taking fields from the request path
(note that this is not yet supported but will be very soon):

- Adds a new `WebhookRequest` class that can be used
  where we need to create and pass around the collection
  of path/method/body/headers.
- Ruby column converts/defaulters accept more parameters.

Additionally, this adds support for creating sequences
tied to integrations in the application database,
so that integrations can use those sequences
when defaulting or converting values.
This acts as a proof of concept for the rest of the system,
since it order to support it we need the service integration
to make it all the way down to the converter.