Sync targets use an advisory lock rather than xaction
Fixes https://github.com/lithictech/webhookdb-api/issues/476

Sync targets would use `SELECT ... FOR UPDATE SKIP LOCKED`
to indicate a row is synced. It would hold this lock for
the duration of the sync, which could be several minutes.
This required a transaction, which also meant that we
couldn't easily partially update the sync target as pages
were synced (just once at the end), as per #570.

Instead, use advisory locks to exclusively sync,
which avoids the need for the transaction.