Refactor sync targets; http supports partial sync
Fixes https://github.com/lithictech/webhookdb-api/issues/570

This is a relatively small functional change
so that we store the latest synced row timestamp
whenever an http sync target processes a page of results
successfully.

However to do it cleanly required a pretty big refactor
of how the sync target code was organized: it previously
was meant for database sync, and http was shimmed in;
now they're equal partners.

This also adds an integration test for HTTP syncing
by opening an TCP server that acts as the HTTP sync target.
So the HTTP target is just another thread in the same process.
Note this means we can't sync in a job (since it'd be in
another process, probably on another server, so isn't reachable),
but that's fine.