Durable jobs keeps track of error details
Durable jobs was never updating the stored job item json.
So we didn't update it after a retry,
nor did we keep track of errors.

Now, we always update the item json whenever we upsert,
and we store error info when we unlock the job
during a job error.

Fixes https://github.com/lithictech/webhookdb-api/issues/606