Fix bug with bulk backfiller
The default `Backfiller#backfill` method wasn't flushing
pending chunks. This isn't actually a responsibility of backfill,
which is why it was missed, but sort of like the `io.ReadCloser`
type of situation in Go, where callers may close the stream
if it is implemented (even if it takes `io.Reader`),
we should flush pending chunks after backfilling if
the backfiller has the right method.

This wasn't hit because none of our bulk backfillers right now
use the default backfill method. But we would have run
into it in the future.