Add support for compound indices
We need compound index support for Google Calendar Events,
since otherwise we end up with a very inefficient filter.
For example, to select a month of events,
we were selecting ALL events for an external id and calendar,
and then filtering by date.

Instead we need to filter all at once.