Better column backfill, fix sponsy days of week
Sponsy made a backwards-incompatible change to their api,
so 'days' is now a string array rather than integer array.

This is also an extremely confusing change because Sponsy
uses the 0th' Day of Week as Monday, not Sunday,
which is more standard.

This modifies the Sponsy 'days' converter to map string names
back onto integers, to preserve the original `days` usage.

It also adds `days_normalized` and `day_names` columns.
These use a new 'backfill statement' and 'backfill expression'
capability used when we `UPDATE` after adding new columns.

Previously, the column/converter would apply the expression
to fill the column from its `data`.

If `backfill_expr` is supplied, it will use that in the
`UPDATE` call, instead of going to the converter
(the converter is too general to use in this case).

It will often be used in conjunction with `backfill_statement`,
which is run **before** the `UPDATE`, and can do something
like creating a temporary function.

This temporary function is required when using `UPDATE`
when trying to map an array column.