Improve how ical handles array properties
Certain properties can occur multiple times,
AND/OR have multiple values within each property.

EXDATE and RDATE are examples of this.

We already handled this with ATTENDEE,
but need to handle more properties.

This change:

- Adds array properties, based on RFC 5545
- Handles RRULE and EXDATE arrays when processing arguments
  for IceCube schedule parameters.

NOTE: This changes the `data` format for this integration,
so is BACKWARDS INCOMPATIBLE with existing client code.
I'm working with Hearth to make sure they can support it,
it'll need to be done before this deploys.