Use custom sidekiq matcher for job side effects
This adds new `have_queue` and `have_empty_queues` matchers
for asserting that jobs are (or are not) enqueued.

This should be used going forward instead of the
`expect(x).to receive(:perform_async).with(...)` mocking form.

It's more reliable and easier to test this way.

Note that you must use `Sidekiq::Testing.fake!` or `sidekiq: :fake`
to use this matcher.