Fix recurring exception bug in google events
There is a bugfix for single events,
that causes a bug for source events.
We end up deleting a source event when a
recurring instance comes through.

Also add and document a way to more easily debug Google Calendar,
by setting up calendar rows via the console rather than having
to reverse engineer all of this when we want to test.