Really handle new Increase ach transfer shape
The update event is missing a ton of fields;
Increase says we should just make an API request for the
updated data rather than expect what's in a webhook.

So we have to handle missing columns,
and also merge the `data` column when we get an update event,
since otherwise the `data` column will be useless.