Handle escaped nulls in #to_json
NULL chars in strings were already removed, but if we had an escaped null char
(`\\\\u0000`) we'd be left with `\\` which resulted in invalid escape sequences.

See https://lithic-technology.sentry.io/issues/4531285265

The NULL ASCII character (\u0000), when present in a string ("\u0000"),
and then encoded into JSON ("\\u0000") is invalid in PG JSONB- its strings cannot contain NULLs
(note that JSONB does not store the encoded string verbatim, it parses it into PG types, and a PG string
cannot contain NULL since C strings are NULL-terminated).

So we remove the "\\u0000" character from encoded JSON- for example, in the hash {x: "\u0000"},
if we #to_json, we end up with '{"x":"\\u0000"}'. The removal of encoded NULL gives us '{"x":""}'.

HOWEVER, if the encoded null is itself escaped, we MUST NOT remove it.
For example, in the hash {x: "\u0000".to_json}.to_json (ie, a JSON string which contains another JSON string),
we end up with '{"x":"\\\\u0000"}`, That is, a string containing the *escaped* null character.
This is valid for PG, because it's not a NULL- it's an escaped "\", followed by "u0000".
If we were to remove the string "\\u0000", we'd end up with '{"x":"\\"}'. This creates an invalid document.

So we remove only "\\u0000" by not replacing "\\\\u0000"- replace all occurences of
"<any one character except backslash>\\u0000" with "<character before backslash>".
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.