Automatic retries on intermittent errors
ServiceBackfiller automatically retries on timeouts,
socket errors, and 5xx, which should resolve some of the flaky errors
in Sentry we see that aren't due to anything we can manage.

This required updating a bunch of 'backfiller retries' tests
that were raising a 5xx error to assert the error was raised.