Issue chunked updates when adding a new column
We need to backfill values into the new column,
but we don't want to lock the entire table
as we update each row. So we need to update in chunks of rows.
Chunk size should be large for speed (and sending over fewer queries),
but small enough to induce a viable delay if another query is updating
the same row.

Note that the delay will only be for writes to those rows;
reads will not block,
so something a bit longer should be ok.

Fixes https://github.com/lithictech/webhookdb-api/issues/634