Add 'demo mode', overhaul README and deploy
Adds a 'demo mode' that can be used for local development
when taking WebhookDB for a spin.

Adds a docker-compose file that uses development mode.

This still needs to be more fully tested,
but it needs to go in so we can do some more testing
and keep moving forward with other work
(this fork needs to get deployed to webhookdb cloud).

This also defaults billing to be disabled by default
(since only webhookdb cloud will be using this).

It also removes some Heroku assumptions around various
deploy mechanisms, like integration tests, though this
is still WIP.