Idempotency: Storage and separate connection
Use `.storage` to store (and retrieve stored) results.

We need this when making external mutation calls while
handling webhooks, like signalwire SMS and Front sync.

Use `.using_seperate_connection` to run the idempotency
(NOT the block) on a separate conn.

Required in some cases, especially when using storage.
Because the result is stored, we can be more granular with
the idempotency call. This is required when
we're in a transaction that we cannot get out of,
like when a lock is required on a row.