Remove non-functional org notification endpoint
It was possible to create a notification (`WebhookSubscription`)
for organizations by leaving out the service integration identifier.
However, we don't yet properly support organization notifications,
so this would result in a busted notification.

This updates the endpoint that `webhookdb notifications create` uses
to prompt if the service integration identifier is not given;
in the future, once org webhooks are supported properly,
we can figure out how to create them on the CLI side.

Fixes https://github.com/webhookdb/webhookdb/issues/775
and closes https://github.com/webhookdb/webhookdb/pull/840