Icalendar: Handle incorrect encoding
We occassionally get incorrectly encoded files.
For example, the response may have a header:

	Content-Type: text/calendar; charset=UTF-8

but the actual encoding is not:

    file -I <filename>
    <filename>: text/calendar; charset=iso-8859-1

In these cases, there's not much we can do.
We can use chardet, but it's a big library and this issue
isn't common enough. Instead, try to force the encoding to utf-8,
which may break some things, but we'll see what happens.