Add support for saved views
Give org admins the ability to manage views in their database
with a new command. It requires a view name and the SELECT query.

This should be safe as long as:

- We validate the view name (as we do for any identifier input)
- We run the SELECT query to verify it is readonly.
- Then we can build our own 'create or replace view' query
  using sanitized input.

However, I want to noodle on this a bit more to make sure
it's fully safe, so it's behind a feature role flag for now.