Backfill job: Do not allow multiple backfills
We were locking a backfill job, but allowing multiple backfill jobs
for the same service integration.

Instead, if a service integration is currently being backfilled,
immediately finish the new job.

Do this with a new lock-specific model,
since we do not want to lock updates to the service integration row.
Otherwise it would block user interactions.