Add Increase OAuth provider
Add a new 'increase app' replicator
and parent the other increase replicators under it.

As for the name 'app',
'auth' wasn't quite right (it relies on events),
and 'oauth' would be ambiguous (because oauth apps are resources
that we can replicate in the future).
'Root' was also imprecise, I think, because we could
allow non-OAuth replicator setup later.

It's possible we want to replace 'intercom_marketplace_root_v1'
and 'front_marketplace_root_v1' with 'app_v1' names,
but since those are found via a marketplace,
rather than through WebhookDB/organic, it seems reasonable
to keep them as-is.