Fix increase deactivate handling
We need to store the Oauth Connection ID for the group, so when there is an oauth disconnect,
we know which replicator to destroy.

The deactivation comes through as an event on the Platform account,
and does not tell us the Group ID. And we cannot fetch the OAuth Connection
on the Platform account, since once a Connection is deactivated,
the Increase API will not return it.

This seems like a bug, and has been reported to Increase.
If their /oauth_connections/:id endpoint starts working for 'inactive' connections,
this code can be removed, and we can look up the group ID when we handle the deactivate webhook.