Overhaul OAuth flow
We were using `/me`-type endpoints to get the OAuth'ed users
email and other info. But we don't want to trust this,
since it requires the oauth provider verifies emails,
which is probably fine, but not something we can depend on.

Instead, always go through the auth flow when authorizing WebhookDB.

Add an 'org chooser' step when doing setup.
We were always creating the replicators in the
first default, admin org, but this is not suitable in
many cases. Someone should be able to create a new org
for their company (not their personal org)
and fill it with replicators, or use an existing org
(as I need to do for Increase).

Also polish up the UI, such as a title and favicon.