Front-signalwire: Handle null SW body
For some unknown reason, signalwire message bodies
can be `null`. If that's the case,
Front errors since `body.body` must be a string.

Use the value `'<no body>'` if the signalwire message body
is null. If it's empty string, we'll still use it as-is,
but I figure it's better to call out the null body
in case it represents something unexpected on the SW side.