Support GET, add Rack Req for sync processing
When processing webhooks synchronously,
we now pass in the `Rack::Request` making the actual request.
This is a new attribute, `Webhookdb::WebhookRequest#rack_request`.

This allows replicators access to the live request,
in case they need to store information like an IP
or request query params.

In theory we could store more of the env/request on `LoggedWebhook`,
but I'm hesitant to add more information there.
Though perhaps it should at least include the query string,
but that can be added when it's needed,
such as if an async replicator needs access to more request info.