Trim whitespace from API inputs
Some shells (one user is reporting iterm2 with zsh)
are adding whitespace, either a new line or space,
at the end of terminal prompts.

I'm not sure why this is, as I cannot reproduce it,
but also, the API should be responsible for trimming whitespace.

This adds a new `TrimmedString` API type that can be used
for any strings sent to the API that should have whitespace stripped
before processing. For example, an httpsync connection url
with trailing whitespace is an invalid URL.

Also modifies `NormalizedEmail` to be a subclass of String
so it can be used directly in Grape's `type:`, rather than
having to use `type: String, coerce_with: NormalizedEmail`.