Developer alerts for backfill errors
We can hit fatal errors during backfills, like revoked keys,
that we need to tell users about using `DeveloperAlert`.

This starts to build out the mechanisms to do so,
starting with the SignalWire Messages replicator.
It's very likely that 1) this will need some more work in the future
and 2) we'll generalize this to other replicators automatically.

But for now, if we get a 401 during SignalWire Messages backfilling,
send a developer alert.

This also adds a class for walking error hierarchies.