Add a 'system log' capability
Makes it much easier to keep a high-level view of what's going on
in the system.

This consists of a new 'system log events' table, viewable in admin,
and some more developer alerts (usually sent to slack) when
something happens.

More alerting can be done through the `ModelEventSystemLogTracker`
job (key off of more events).