Fix front signalwire message edge case
We were replacing the `data` column with the full row,
which was giving us a `data` shape we weren't expecting,
causing errors during backfill (on receipt of a signalwire message).

This would only manifest under certain conditions which kicked off
a backfill in this situation, I'm not entirely sure why.
But it was definitely a bug which needed to be fixed.