Add test and docs for explicit NULL backfill
In some cases, we need to use a `backfill_expr`
to explicitly set `NULL` rather than an auto-generated default
expression (trying to derive the value from `data`).

This is a little confusing because you need to use
`backfill_expr: Sequel[nil]` rather than `backfill_expr: nil`,
so I documented and tested it explicitly.