ICalendar: Add sync time, event count, feed size
Some ICalendar feeds are very large, which cause performance problems.
Large feed files take a while to download,
and feeds with many events take a while to upsert
(a big feed with few recurring events can still be fast,
and a small feed with many recurring events can be slow).

Start tracking stats about the feed when we sync,
so we can use it to dynamically schedule-
small/fast feeds can be reprocessed often and slower feeds
can be processed less often.