Use optimized form of 'exclude(col => values)'
WIP: Actually I think the perf improvements were due to hitting shared caches,
this doesn't work.

Add and use `Dbutil.where_not_in_using_index`.

Fixes slow ical 'set all not-seen rows to canceled' performance.

Faster version of dataset.exclude(column => values)
when +column+ has an index.
Instead of "where column not in values", which cannot use the index on +column+
(it has to walk each row to check if the value is in +values+),
it uses, "where +column+ not in the set of rows that have a column in +values+".
Concretely, instead of:

  SELECT * FROM mytable
  WHERE mycolumn NOT IN ('a', 'b', 'c')

use this method to build the query:

  SELECT * FROM mytable
  WHERE mycolumn NOT IN (SELECT mycolumn FROM mytable WHERE mycolumn IN ('a', 'b', 'c'))

The explain plan of the first includes:
  Filter: (mycolumn <> ALL ('{a,b,c}'::text[]))

The explain plan of the second instead includes something like:
  Filter: (NOT (hashed SubPlan 1))
    SubPlan 1
     ->  Index Only Scan using mycolumn_idx on mytable
           Index Cond: (mycolumn = ANY ('{a,b,c}'::text[]))

I'm not entirely sure why this makes such a difference,
but especially with large lists, it certainly does.
For smaller lists it may not make a difference, or may have a small negative impact.
Postgres17 may obvious the need for this due to IN changes, if so it can be removed.