Re-implement stale row deleter for massive tables
Running the stale row deleter on massive (billion-row?)
tables could cause problems.
This uses a new algorithm to issue a series of small deletes
by walking the delete window in hour chunks,
and avoiding a sequential scan and also allowing faster
re-deletes because we don't have to walk a larger index space
for each chunk.