Improve icalendar sync enqueue logic
We are enqueing too many icalendar jobs due to incorrect
reasoning in the enqueue job, and the sync job will
happily reprocess too often.

This change:

- Causes the sync job to noop if the calendar has been recently synced.
- Reworks the enqueue logic.
  See the job doc for an explanation/reasoning of how it works,
  but in brief, it runs every hour and enqueues all rows
  needing a sync for that next hour.

With this change, we should see a tighter sync timing,
so we'll resync every 10 hours rather than 6.

Also adjust the minimum event recurrence to the year 2000,
rather than 1990, I don't think the 10 years is useful
considering when we started to use the internet
and digital calendars...