Fix long transactions with sync targets
Running an httpsync uses a serverside cursor,
which opens a transaction.
The transaction can be open a long time, which will block dead tuples
being cleaned up across the entire database.
That is, a slow httpsync (which is expected) can block
any of the entire day's tuples being cleaned up until
the job is restarted.

This adds a `SYNC_TARGET_MAX_TRANSACTION_SECONDS` setting
which will re-enqueue the sync target if it runs
more than the configured amount of time.
This ensures we'll only see sync jobs running for 10 minutes,
rather than 24 hours, and should see tuples reclaimed correctly.