Upsert multiple rows for one call
`_resource_and_event` can now return an array of resources
which are each inserted individually.
This preserves current semantics around upserting,
but allows replicators to take one 'row' and derive
many from it.

Note that this is only possible to use when upserting,
and when the event part of resource_and_event is nil.

In the future, we could make it possible to support
multiple rows with upsert:false, but since upsert:false
assumes we use the returned value, and an array of resources
changes the return type from a hash to an array,
this seems risky and hard to program against.