Front/Signalwire: Support multiple recipients
We were inserting a front/signalwire channel message only for
the first recipient, so multi-message blasts did not work.

Instead, insert a row for each recipient.
This requires setting an external_id that is a combo
of a message ID and recipient.

Also the returned external_id from the front webhook
is now a 'fake' id based on all the message external ids.
Not sure if we can do anything better than this.