Icalendar: Pre-check for feed changes
Enqueuing icalendar sync jobs now checks in advance if
the feed has likely changed. It does this in bulk using
a threadpool. It introduces extra work in the worst case
(many changes) but eliminates a lot of redundant work,
especially for more rapid sync timing.

That is, resyncing feeds every hour would be impossible at scale
due to the database churn; but if a large fraction of feeds
have not changed (which is usually the case), we can eliminate
trying to sync them for the cost of an HTTP request and minimal
processing on the Ruby side.

This checking comes at a significant clocktime penalty,
but we can largely eliminate that through parallelization,
since most of the 'work' is in HTTP requests.